+++ Assets/Resources/instructionList.json
@@ -8,7 +8,7 @@
     "driving": [
         "Use WASD to control your car.",
         "Holding (left) Shift enables the handbrake.",
-        "Try clearing this course to proceed.",
+        "To make sure you understand, try clearing this course to proceed.",
         "Looks good."
     ],
