--- Assets/minimapDisable.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 09d5348822d5db04caf5b5b6f6b69534
