+++ Assets/scripts/camerathing.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9ab82409649ee254e80c7e7e01ee3a58
