+++ Assets/scripts/minimapDisable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 09d5348822d5db04caf5b5b6f6b69534
