--- Assets/sound/music/tutorialTracks/5_beforedrift_MAINloop.wav.meta
@@ -1,23 +1,0 @@
-fileFormatVersion: 2
-guid: b84813d1d8dfeb74bb352ecd8b5fa781
-AudioImporter:
-  externalObjects: {}
-  serializedVersion: 8
-  defaultSettings:
-    serializedVersion: 2
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-    preloadAudioData: 0
-  platformSettingOverrides: {}
-  forceToMono: 0
-  normalize: 1
-  loadInBackground: 0
-  ambisonic: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
