+++ Assets/sound/music/tutorialTracks/7_FINAL_TUTORIAL_main.wav.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 43e6faa3c27df7748aef3b1c59ac2a9f
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 8
+  defaultSettings:
+    serializedVersion: 2
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+    preloadAudioData: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
