+++ Assets/sound/music/tutorialTracks/7_FINAL_TUTORIAL_turbe.wav.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 0c99aa3c7c0d5ac4480ab3e56911682e
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 8
+  defaultSettings:
+    serializedVersion: 2
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+    preloadAudioData: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
