--- Assets/sound/music/tutorialTracks/7_turbe_DRIFTloop_fix.wav.meta
@@ -1,23 +1,0 @@
-fileFormatVersion: 2
-guid: 9a6f0ba02792a074f8327b5fd12dbf26
-AudioImporter:
-  externalObjects: {}
-  serializedVersion: 8
-  defaultSettings:
-    serializedVersion: 2
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-    preloadAudioData: 0
-  platformSettingOverrides: {}
-  forceToMono: 0
-  normalize: 1
-  loadInBackground: 0
-  ambisonic: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
